add npc-utils-js.js.
